+++ packages/experimental/Checkbox/CHANGELOG.md
@@ -1,8 +1,102 @@
 # Change Log - @fluentui-react-native/experimental-checkbox
 
-This log was last generated on Mon, 28 Nov 2022 11:51:06 GMT and should not be manually modified.
+This log was last generated on Tue, 27 Dec 2022 10:18:14 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.16
+
+Tue, 27 Dec 2022 10:18:14 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.6
+- Bump @fluentui-react-native/framework to v0.8.27
+
+## 0.13.15
+
+Fri, 23 Dec 2022 11:03:02 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.5
+
+## 0.13.14
+
+Wed, 21 Dec 2022 06:39:34 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.4
+
+## 0.13.13
+
+Thu, 15 Dec 2022 14:50:50 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.3
+- Bump @fluentui-react-native/framework to v0.8.26
+
+## 0.13.12
+
+Tue, 13 Dec 2022 23:18:04 GMT
+
+### Patches
+
+- Fix typos (ruaraki@microsoft.com)
+
+## 0.13.11
+
+Fri, 09 Dec 2022 05:23:38 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.2
+- Bump @fluentui-react-native/framework to v0.8.25
+
+## 0.13.10
+
+Tue, 06 Dec 2022 16:59:54 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.1
+- Bump @fluentui-react-native/framework to v0.8.24
+
+## 0.13.9
+
+Tue, 06 Dec 2022 02:02:55 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.18.0
+
+## 0.13.8
+
+Tue, 06 Dec 2022 00:46:41 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.17.8
+- Bump @fluentui-react-native/framework to v0.8.23
+
+## 0.13.7
+
+Thu, 01 Dec 2022 03:12:20 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.17.7
+- Bump @fluentui-react-native/framework to v0.8.22
+
+## 0.13.6
+
+Mon, 28 Nov 2022 19:49:35 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.17.6
 
 ## 0.13.5