+++ packages/experimental/NativeFontMetrics/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/experimental-native-font-metrics
 
-This log was last generated on Mon, 07 Nov 2022 21:42:58 GMT and should not be manually modified.
+This log was last generated on Mon, 28 Nov 2022 19:49:35 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.0
+
+Mon, 28 Nov 2022 19:49:35 GMT
+
+### Minor changes
+
+- Send new font metrics information through a JS event (adgleitm@microsoft.com)
 
 ## 0.2.0